Dinky5: lose regfile maintenance rules
This makes the design significantly simpler by making internal buses
implicit in method calls. I'm not entirely sure why I didn't do this
earlier, but I think it had to do with the RegFile not being
encapsulated.

This makes the SoC synthesis results slightly larger, which is
disappointing, but the reduction in source code and moving parts appeals
to me.