Twisty5: pipeline comparator to break cp
This change moves 23 bits of the comparator into a new pipeline stage in
between previous stages 1 and 2. 1118 LCs.

I'm not totally convinced this is a good idea.

Positive:

- Boosts clock frequency to 78MHz.

Negative:

- Costs 162 LCs, which is basically an entire bidirectional barrel
  shifter. Leaves only 162 LCs free for, well, anything else.
- Additional complexity, obvs
- By increasing pipeline depth, this eats the remaining cycle of latency
  we had for increasing bus slack. It is no longer possible to put a
  FIFO between the CPU and RAM without bypassing cache in stage1.

I think I'll see how it goes; the fact that this boosts each hart's
performance to approximately the throughput of the other cores I've
written is pretty compelling. (As long as you don't need to shift
anything, of course.)