Twisty5: remove serial shifter
The serial shifter had an implementation bug and a concept flaw and does
not appear to have ever worked. I have removed it in favor of the
combinational barrel shifter, effectively removing the small/medium
variants of the SoC.

We are now very cramped for space in the hx1k (92% used) and fmax is
dropping as routing overhead increases.

On the up side, knowing the barrel shifter will always be present in
stage3 means I can use it in other places in stage3. I have replaced the
load datapath's bespoke right-byte-shifter with the barrel shifter for a
small area savings.