Twisty5: rotate bus port to stages 4-1
This drives the basic instruction decode in stage 1 directly from the
bus response lines, which is fine, because it's trivial. In exchange
this takes the address decoding out of stage 3, slightly shortening the
critical path.

And it's slightly smaller now, for reasons I can't explain -- we're
close enough to filling the chip that the router is doing random stuff.