Twisty5: balance stage2/3 by moving barrel shifter
After banging my head against the cryptic critical path report coming
out of yosys/nextpnr, I eventually worked out that it was the
load-shift-extend datapath that was slowing us down. By moving the
shift part into stage 2 (and keeping the extend part in stage 3) I've
recovered a smidge over 33% more performance.

The critical path is back to being easily recognizable as the top end of
the comparator and conditional branch logic, so that's nice.